Issue #34: Fix range limitation
-Range now use long instead of int to allow larger values.
-Fix issue in test when not using ECD.
-Fix typo.

+++ src/main/java/io/openio/sds/models/Range.java
@@ -9,35 +9,30 @@ import java.util.regex.Pattern;
 
 import io.openio.sds.common.MoreObjects;
 
-/**
- * 
- *
- *
- */
 public class Range {
 
     private static final Pattern RANGE_PATTERN = Pattern
             .compile("^([\\d]+)?-([\\d]+)?$");
 
-    private int from = 0;
-    private int to = -1;
+    private long from = 0;
+    private long to = -1;
 
-    private Range(int from, int to) {
+    private Range(long from, long to) {
         this.from = from;
         this.to = to;
     }
 
-    public static Range upTo(int to) {
+    public static Range upTo(long to) {
         checkArgument(0 < to);
         return new Range(0, to);
     }
 
-    public static Range from(int from) {
+    public static Range from(long from) {
         checkArgument(0 <= from);
         return new Range(from, -1);
     }
 
-    public static Range between(int from, int to) {
+    public static Range between(long from, long to) {
         checkArgument(from >= 0 && to > 0 && to >= from,
                 "Invalid range");
         return new Range(from, to);
@@ -54,11 +49,11 @@ public class Range {
                 : between(parseInt(m.group(1)), parseInt(m.group(2)));
     }
 
-    public int from() {
+    public long from() {
         return from;
     }
 
-    public int to() {
+    public long to() {
         return to;
     }
